backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.6
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.3.6
eventlet==0.36.1
requests==2.32.3
python-nmap==0.7.1
dnspython==2.7.0
scapy==2.5.0
rq>=1.15.0
python-dotenv>=1.0.0
pdfkit==1.0.0
pyOpenSSL==24.0.0
aiohttp==3.9.3
shodan==1.31.0
censys==2.2.12
paramiko==3.4.0
xmltodict==0.13.0
Jinja2==3.1.3
PyYAML==6.0.1



